+++ src/loginServices.test.js
@@ -271,11 +271,13 @@ describe('validateUser', () => {
     const store = {
       dispatch: jest.fn(),
     };
+    const handleError = jest.fn().mockReturnValue(Promise.resolve());
 
     mockFetchError();
 
-    await validateUser('url', store, 'tenant', {});
-    expect(store.dispatch).toHaveBeenCalledWith(setServerDown());
+    const res = await validateUser('url', store, 'tenant', {}, handleError);
+    expect(handleError).toHaveBeenCalled();
+    expect(res).toBeUndefined();
     mockFetchCleanUp();
   });
 
@@ -385,14 +387,15 @@ describe('validateUser', () => {
       dispatch: jest.fn(),
       getState: () => ({ okapi: { tenant: 'monkey' } }),
     };
+    const handleError = jest.fn().mockReturnValue(Promise.resolve());
 
     global.fetch = jest.fn().mockImplementation(() => {
-      return Promise.resolve({ ok: false });
+      return Promise.resolve({ ok: false, text: () => Promise.resolve('boom') });
     });
 
-    await validateUser('url', store, 'tenant', {});
-    expect(store.dispatch).toHaveBeenCalledWith(clearCurrentUser());
-    expect(store.dispatch).toHaveBeenCalledWith(setServerDown());
+    const res = await validateUser('url', store, 'tenant', {}, handleError);
+    expect(handleError).toHaveBeenCalled();
+    expect(res).toBeUndefined();
     mockFetchCleanUp();
   });
 });